Phase 6: Complete remaining engine gaps — editor panels, serialization, rendering, audio DSP
- Implement AssetBrowserPanel::on_render() and complete all editor panel stubs
- Create game_state_serializer.cpp for save/load game state as JSON
- Add TilemapComponent serialization to scene_serializer.cpp
- Implement 2D frustum culling, tilemap rendering, and glyph rendering in BatchRenderer2D
- Add shadow mapping integration to ForwardRenderer3D
- Implement audio DSP effects (reverb, low-pass, high-pass filters)
- Add bitmap font text rendering to UI system
- Create debug overlay for profiler stats display
- Add Profiler and LODEvaluator integration to runtime main loop
- Fix build: add bitmap_font.cpp to UI CMake, link Nexus::Perf in runtime
- Fix warnings: ReverbEffect constructor, AudioBus initializers, type conversions

All 865 tests pass, clean build with no warnings.

+++ engine/ui/src/bitmap_font.cpp
@@ -192,15 +192,15 @@ bool BitmapFont::load_bmfont(const std::string& fnt_path) {
             while (iss >> token) {
                 if (token.substr(0, 11) == "lineHeight=") line_height_ = std::stof(token.substr(11));
                 else if (token.substr(0, 5) == "base=") base_line_ = std::stof(token.substr(5));
-                else if (token.substr(0, 7) == "scaleW=") atlas_w = std::stoul(token.substr(7));
-                else if (token.substr(0, 7) == "scaleH=") atlas_h = std::stoul(token.substr(7));
+                else if (token.substr(0, 7) == "scaleW=") atlas_w = static_cast<u32>(std::stoul(token.substr(7)));
+                else if (token.substr(0, 7) == "scaleH=") atlas_h = static_cast<u32>(std::stoul(token.substr(7)));
             }
         } else if (tag == "char") {
             Glyph g{};
             float x = 0, y = 0, w = 0, h = 0;
             std::string token;
             while (iss >> token) {
-                if (token.substr(0, 3) == "id=") g.codepoint = std::stoul(token.substr(3));
+                if (token.substr(0, 3) == "id=") g.codepoint = static_cast<u32>(std::stoul(token.substr(3)));
                 else if (token.substr(0, 2) == "x=") x = std::stof(token.substr(2));
                 else if (token.substr(0, 2) == "y=") y = std::stof(token.substr(2));
                 else if (token.substr(0, 6) == "width=") w = std::stof(token.substr(6));
